Add error code to unrecognised error message
From time to time, we get an error that is not recognised by our
messages object, in which case, only the message is returned.

This means that if we want to add a new message rule for that specific
type of error, we must reproduce the error with `DEBUG=true` to get it's
code. Also notice that some one-time errors are very hard to reproduce.

For this reason, this module now returns the error code along with the
error message when the error is not recognised.